CI: Add a restoration key for Azure's remote data caches
We do a similar trick for the compiler cache. This allows each builder
to separately push their local data cache (if it changed) while pulling
a shared cache, without the race outlined in commit a10f869. This is
needed for a subsequent commit which will enable this cache for Fuzzer
builds.